+++ pkg/cluster/internal/create/actions/createworker/provider.go
@@ -191,7 +191,7 @@ func installCalico(n nodes.Node, k string, keosCluster commons.KeosCluster, allo
 	}
 
 	// Wait for calico-system namespace to be created
-	c = "timeout 30s bash -c 'until kubectl --kubeconfig " + kubeconfigPath + " get ns calico-system; do sleep 2s ; done'"
+	c = "timeout 60s bash -c 'until kubectl --kubeconfig " + kubeconfigPath + " get ns calico-system; do sleep 2s ; done'"
 	_, err = commons.ExecuteCommand(n, c)
 	if err != nil {
 		return errors.Wrap(err, "failed to wait for calico-system namespace")