+++ pkg/cmd/kind/create/cluster/validation/awsvalidator.go
@@ -17,43 +17,48 @@ import (
 	"sigs.k8s.io/kind/pkg/commons"
 )
 
-var eksInstance *EKSValidator
+var awsInstance *AWSValidator
+var provisionersTypesAWS = []string{"io1", "io2", "gp2", "gp3", "sc1", "st1", "standard", "sbp1", "sbg1"}
 
 const (
 	cidrSizeMax = 65536
 	cidrSizeMin = 16
 )
 
-type EKSValidator struct {
+type AWSValidator struct {
 	commonValidator
+	managed bool
 }
 
-func newEKSValidator() *EKSValidator {
-	if eksInstance == nil {
-		eksInstance = new(EKSValidator)
+func newAWSValidator(managed bool) *AWSValidator {
+	if awsInstance == nil {
+		awsInstance = new(AWSValidator)
 	}
-	return eksInstance
+	awsInstance.managed = managed
+	return awsInstance
 }
 
-var provisionersTypesAWS = []string{"io1", "io2", "gp2", "gp3", "sc1", "st1", "standard", "sbp1", "sbg1"}
-
-func (v *EKSValidator) Spec(spec commons.Spec) {
+func (v *AWSValidator) Spec(spec commons.Spec) {
 	v.descriptor = spec
 }
 
-func (v *EKSValidator) SecretsFile(secrets commons.SecretsFile) {
+func (v *AWSValidator) DescriptorFile(spec commons.Spec) {
+	v.descriptor = spec
+}
+
+func (v *AWSValidator) SecretsFile(secrets commons.SecretsFile) {
 	v.secrets = secrets
 }
 
-func (v *EKSValidator) Validate(fileType string) error {
+func (v *AWSValidator) Validate(fileType string) error {
 	switch fileType {
 	case "descriptor":
-		err := v.descriptorEksValidations((*v).descriptor, (*v).secrets)
+		err := v.descriptorAwsValidations((*v).descriptor, (*v).secrets)
 		if err != nil {
 			return err
 		}
 	case "secrets":
-		err := secretsEksValidations((*v).secrets)
+		err := secretsAwsValidations((*v).secrets)
 		if err != nil {
 			return err
 		}
@@ -63,7 +68,7 @@ func (v *EKSValidator) Validate(fileType string) error {
 	return nil
 }
 
-func (v *EKSValidator) CommonsValidations() error {
+func (v *AWSValidator) CommonsValidations() error {
 	err := commonsValidations((*v).descriptor, (*v).secrets)
 	if err != nil {
 		return err
@@ -71,7 +76,7 @@ func (v *EKSValidator) CommonsValidations() error {
 	return nil
 }
 
-func (v *EKSValidator) descriptorEksValidations(spec commons.Spec, secretsFile commons.SecretsFile) error {
+func (v *AWSValidator) descriptorAwsValidations(spec commons.Spec, secretsFile commons.SecretsFile) error {
 	err := commonsDescriptorValidation(spec)
 	if err != nil {
 		return err
@@ -80,9 +85,11 @@ func (v *EKSValidator) descriptorEksValidations(spec commons.Spec, secretsFile c
 	if err != nil {
 		return err
 	}
-	err = eksAZValidation(spec, secretsFile)
-	if err != nil {
-		return err
+	if spec.ControlPlane.Managed {
+		err = eksAZValidation(spec, secretsFile)
+		if err != nil {
+			return err
+		}
 	}
 	err = v.storageClassValidation(spec)
 	if err != nil {
@@ -91,7 +98,7 @@ func (v *EKSValidator) descriptorEksValidations(spec commons.Spec, secretsFile c
 	return nil
 }
 
-func secretsEksValidations(secretsFile commons.SecretsFile) error {
+func secretsAwsValidations(secretsFile commons.SecretsFile) error {
 	err := commonsSecretsValidations(secretsFile)
 	if err != nil {
 		return err
@@ -225,7 +232,7 @@ func filterPrivateSubnet(svc *ec2.EC2, subnetID *string) (string, error) {
 	}
 }
 
-func (v *EKSValidator) storageClassValidation(spec commons.Spec) error {
+func (v *AWSValidator) storageClassValidation(spec commons.Spec) error {
 	if spec.StorageClass.EncryptionKey != "" {
 		err := v.storageClassKeyFormatValidation(spec.StorageClass.EncryptionKey)
 		if err != nil {
@@ -240,7 +247,7 @@ func (v *EKSValidator) storageClassValidation(spec commons.Spec) error {
 	return nil
 }
 
-func (v *EKSValidator) storageClassKeyFormatValidation(key string) error {
+func (v *AWSValidator) storageClassKeyFormatValidation(key string) error {
 	regex := regexp.MustCompile(`^arn:aws:kms:[a-zA-Z0-9-]+:\d{12}:key/[a-zA-Z0-9-_]+$`)
 	if !regex.MatchString(key) {
 		return errors.New("Incorrect key for encryption format. It must have the complete arn format")
@@ -248,7 +255,7 @@ func (v *EKSValidator) storageClassKeyFormatValidation(key string) error {
 	return nil
 }
 
-func (v *EKSValidator) storageClassParametersValidation(spec commons.Spec) error {
+func (v *AWSValidator) storageClassParametersValidation(spec commons.Spec) error {
 	sc := spec.StorageClass
 	typesSupportedForIOPS := []string{"io1", "io2", "gp3"}
 	fstypes := []string{"xfs", "ext3", "ext4", "ext2"}