+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 * [Azure] Add Azure file CSI driver
 * [Azure] Bump cluster-api-provider-azure to v1.10.0: Fix Azure VMs load balancer health check
 * [GCP] Bump cluster-api-provider-gcp to v1.4.0: Fix GCP VMs load balancer health check
+* [Core] Bump cluster api to v1.5.1
+* [Core] Bump Calico to v3.26.1
+* [AWS] Bump cluster-api-provider-aws to v1.5.1
+* [AWS] Bump clusterawsadm to v2.2.1
 
 ## 0.17.0-0.2.0 (2023-07-03)